requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-rest-registration==0.7.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.10
whitenoise==6.0.0
